requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
redis>=5.2.0
python-dotenv>=1.0.0
pydantic>=2.10.0
google-generativeai>=0.8.0
aiohttp>=3.11.0
cachetools>=5.5.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
